WIP type cleanups
